+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 ### Changed
 ### Fixed
+- DF-1077 Fixed support for UUID and Timestamp data types
 
 ## [0.5.0] - 2017-03-03
 - Major restructuring to upgrade to Laravel 5.4 and be more dynamically available